merged the pick and setup logic into one.
I merged the pick and setup logic since trying to implement anything else felt bad. Feels akward t have a big master class like this, but honestly it's for the best. Tried to implement an observer pattern, but it was just bad and not the right use case. Giant logic hell hole is probably best. Updated tests and names accordingly.